New target --dist-software to package all necessary software tarballs
When publishing a project, it is necessary to also publish the source code
of all necessary software of the project. We had recently added a new
'./project make' target called 'dist-software' for this job, but had
forgotten to add it in the output of './project --help'! There was also a
small bug inside of it that didn't allow the successful copying of the
created tarball to the top project directory.

With this commit, an explanation for this target has been added in the
output of './project --help' and that bug has been fixed.