Removing preparation-done.mk when cleaning by ./project make clean
Until this commit, the file `BDIR/software/preparation-done.mk' were not
removed when cleaning the project with `./project make clean'. This file
is generated in the preparation of the data during the analysis step.
However, the cleaning is expected to remove anything generated in the
analysis process! Step by step, with the commands:

./project make        --->  Will make the preparation and analysis
./project make clean  --->  Will remove all analysis outputs (but
                            not `preparation-done.mk')
./project make        --->  Won't do the preparation, only analysis!

However, in the last step it should do the preparation again, because
the input data could have change for any reason. With this commit, the
file `BDIR/software/preparation-done.mk' is removed when cleaning the
project, and consequently, in the analysis step the input data is
prepared.