The distclean target accounts for non-existance of git hooks
Until now, when the user wanted to complete remove all built files
(including software), the './project make distclean' command would fail if
the git hooks weren't installed. They are present when the project's
configuration has been successfully finished, but this bug can happen when
trying to re-do an incomplete build.

With this commit, this is fixed by adding an '-f' has been added before the
'rm' command for the Git hooks.