Entered data and software directories stored as absolute addresses
Until now, when the user specified an input and software directory, the raw
string they entered was used. But when this string was a relative location,
this could be problematic in general scenarios.

With this commit, the same function that finds the absolute location of the
build directory is used to find the absolute address of the data and
software directories.