README-hacking.md: added new paper using Maneage (arXiv:2007.11779)
Roukema+2020 (arXiv:2007.11779) is a newly published (as preprint) paper
that uses Maneage, so it is being added to the list of published or
submitted papers in 'README-hacking.md'. The Software Heritage URL sticks
out way beyond the standard number of columns in the plain text form of the
updated 'README-hacking.md' file, when rendered using markdown, it
shouldn't look so bad.

Also, see the related task https://savannah.nongnu.org/task/index.php?15736
(Raul+2020 should be Infante-Sainz+2020) for a suggestion of a more
standard machine-readable format.

It should be mentioned and emphasised to the reader that one should very
carefully and obediently note and pay attention to the noteworthy fact that
a few distracting words [1] such as "Note that" are removed in this
commit. ;)

[1] https://en.wiktionary.org/wiki/pontification