README-hacking.md: updated paper to cite for using Maneage
Until now, we were asking the users of Maneage to cite the first paper that
used its primoridal version (arXiv:1505:01664). But there is now a paper
that fully describes the concept (arXiv:2006.03018).

With this commit, in the 'citation' section of 'README-hacking.md' we now
ask to cite the new paper.