IMPORTANT: organizational improvements in Maneage TeX sources
This only concerns the TeX sources in the default branch. In case you don't
use them, there should only be a clean conflict in 'paper.tex' (that is
obvious and easy to fix). Conflicts may only happen in some of the
'tex/src/preamble-*.tex' files if you have actually changed them for your
project. But generally any conflict that does arise by this commit with
your project branch should be very clear and easy to fix and test.

In short, from now on things will even be easier: any LaTeX configuration
that you want to do for your project can be done in
'tex/src/preamble-project.tex', so you don't have to worry about any other
LaTeX preamble file. They are either templates (like the ones for PGFPlots
and BibLaTeX) or low-level things directly related to Maneage. Until now,
this distinction wasn't too clear.

Here is a summary of the improvements:

 - Two new options to './project make': with '--highlight-new' and
   '--highlight-notes' it is now possible to activate highlighting on the
   command-line. Until now, there was a LaTeX macro for this at the start
   of 'paper.tex' (\highlightchanges). But changing that line would change
   the Git commit hash, making it hard for the readers to trust that this
   is the same PDF. With these two new run-time options, the printed commit
   hash will not changed.

 - paper.tex: the sentences are formatted as one sentence per line (and one
   line per sentence). This helps in version controlling narrative and
   following the changes per sentence. A description of this format (and
   its advantages) is also included in the default text.

 - The internal Maneage preambles have been modified:

    - 'tex/src/preamble-header.tex' and 'tex/src/preamble-style.tex' have
      been merged into one preamble file called
      'tex/src/preamble-maneage-default-style.tex'. This helps a lot in
      simply removing it when you use a journal style file for example.

    - Things like the options to highlight parts of the text are now put in
      a special 'tex/src/preamble-maneage.tex'. This helps highlight that
      these are Maneage-specific features that are independent of the style
      used in the paper.

    - There is a new 'tex/src/preamble-project.tex' that is the place you
      can add your project-specific customizations.