Configuration: GNU Binutils linking bug on some systems fixed
Until now, when building GNU Binutils on GNU Linux operating systems, we
would simply put a link to the host's core C library components (the
'*crt*' files). However, the symbolic link wasn't "forced"! So if it
already existed in the build directory, it would crash.

With this commit a '-f' option has been added to the 'ln' command and this
fixed the problem.

This bug was reported by Zahra Sharbaf.